test: fold dispatcher-error tests into lint suites
Move the four dispatcher-guard assertions out of the scratch
`test/valkey/test_dispatcher_errors.rb` file into the matching
`test/lint/*_commands.rb` modules so they run alongside the rest of
each command family and are exercised in both the standalone and
cluster suites (Lint modules are mixed into both runners).

- BB-3-004 (`script(:garbage)`) → test/lint/scripting_commands.rb
- BB-4-005 (`client(:garbage)`, :getname/:setname parity) → test/lint/connection_commands.rb
- BB-4-006 (`cluster(:garbage)`) → test/lint/cluster_commands.rb

Remove the scratch file and the standalone-only wire-in.
